Move to common/stages
